+++ settings.php
@@ -26,8 +26,17 @@
  */
 
 defined('MOODLE_INTERNAL') || die;
+$settings = null;
+if (is_siteadmin()) {
 
-if ($ADMIN->fulltree) {
+    $ADMIN->add('themes', new admin_category('theme_shoelace', 'Shoelace'));
+
+    // General settings.
+    $temp = new admin_settingpage('theme_shoelace_generic', get_string('generalsettings', 'theme_shoelace'));
+
+    $temp->add(new admin_setting_heading('theme_shoelace_generalheading',
+        get_string('generalheadingsub', 'theme_shoelace'),
+        format_text(get_string('generalheadingdesc', 'theme_shoelace'), FORMAT_MARKDOWN)));
 
     // Theme colour setting.
     $name = 'theme_shoelace/themecolour';
@@ -37,7 +46,7 @@ if ($ADMIN->fulltree) {
     $previewconfig = null;
     $setting = new admin_setting_configcolourpicker($name, $title, $description, $default, $previewconfig);
     $setting->set_updatedcallback('theme_reset_all_caches');
-    $settings->add($setting);
+    $temp->add($setting);
 
     // Text colour setting.
     $name = 'theme_shoelace/textcolour';
@@ -47,7 +56,7 @@ if ($ADMIN->fulltree) {
     $previewconfig = null;
     $setting = new admin_setting_configcolourpicker($name, $title, $description, $default, $previewconfig);
     $setting->set_updatedcallback('theme_reset_all_caches');
-    $settings->add($setting);
+    $temp->add($setting);
 
     // Navbar text colour setting.
     $name = 'theme_shoelace/navbartextcolour';
@@ -57,7 +66,7 @@ if ($ADMIN->fulltree) {
     $previewconfig = null;
     $setting = new admin_setting_configcolourpicker($name, $title, $description, $default, $previewconfig);
     $setting->set_updatedcallback('theme_reset_all_caches');
-    $settings->add($setting);
+    $temp->add($setting);
 
     /* CDN Fonts - 1 = no, 2 = yes. */
     $name = 'theme_shoelace/cdnfonts';
@@ -68,7 +77,7 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'),   // No.
         2 => new lang_string('yes')   // Yes.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $temp->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     // Invert Navbar to dark background.
     $name = 'theme_shoelace/invert';
@@ -76,7 +85,7 @@ if ($ADMIN->fulltree) {
     $description = get_string('invertdesc', 'theme_shoelace');
     $setting = new admin_setting_configcheckbox($name, $title, $description, 0);
     $setting->set_updatedcallback('theme_reset_all_caches');
-    $settings->add($setting);
+    $temp->add($setting);
 
     // Logo file setting.
     $name = 'theme_shoelace/logo';
@@ -84,7 +93,7 @@ if ($ADMIN->fulltree) {
     $description = get_string('logodesc', 'theme_shoelace');
     $setting = new admin_setting_configstoredfile($name, $title, $description, 'logo');
     $setting->set_updatedcallback('theme_reset_all_caches');
-    $settings->add($setting);
+    $temp->add($setting);
 
     // Number of marketing blocks.
     $name = 'theme_shoelace/nummarketingblocks';
@@ -98,7 +107,7 @@ if ($ADMIN->fulltree) {
     );
     $default = 2;
     $setting = new admin_setting_configselect($name, $title, $description, $default, $choices);
-    $settings->add($setting);
+    $temp->add($setting);
 
     // Number of footer blocks.
     $name = 'theme_shoelace/numfooterblocks';
@@ -112,7 +121,7 @@ if ($ADMIN->fulltree) {
     );
     $default = 2;
     $setting = new admin_setting_configselect($name, $title, $description, $default, $choices);
-    $settings->add($setting);
+    $temp->add($setting);
 
     // Footnote setting.
     $name = 'theme_shoelace/footnote';
@@ -121,7 +130,7 @@ if ($ADMIN->fulltree) {
     $default = '';
     $setting = new admin_setting_confightmleditor($name, $title, $description, $default);
     $setting->set_updatedcallback('theme_reset_all_caches');
-    $settings->add($setting);
+    $temp->add($setting);
 
     // Custom CSS file.
     $name = 'theme_shoelace/customcss';
@@ -130,5 +139,33 @@ if ($ADMIN->fulltree) {
     $default = '';
     $setting = new admin_setting_configtextarea($name, $title, $description, $default);
     $setting->set_updatedcallback('theme_reset_all_caches');
-    $settings->add($setting);
+    $temp->add($setting);
+
+    $ADMIN->add('theme_shoelace', $temp);
+
+    // Style guide.
+    if (file_exists("{$CFG->dirroot}/theme/shoelace/shoelace_admin_setting_styleguide.php")) {
+        require_once($CFG->dirroot . '/theme/shoelace/shoelace_admin_setting_styleguide.php');
+    } else if (!empty($CFG->themedir) && file_exists("{$CFG->themedir}/shoelace/shoelace_admin_setting_styleguide.php")) {
+        require_once($CFG->themedir . '/shoelace/shoelace_admin_setting_styleguide.php');
+    }
+    $temp = new admin_settingpage('theme_shoelace_styleguide', get_string('styleguide', 'theme_shoelace'));
+    $temp->add(new shoelace_admin_setting_styleguide('theme_shoelace_styleguide',
+        get_string('styleguidesub', 'theme_shoelace'),
+        get_string('styleguidedesc', 'theme_shoelace',
+            array(
+                'origcodelicenseurl' => html_writer::link('http://www.apache.org/licenses/LICENSE-2.0', 'Apache License v2.0',
+                    array('target' => '_blank')),
+                'thiscodelicenseurl' => html_writer::link('http://www.gnu.org/copyleft/gpl.html', 'GPLv3',
+                    array('target' => '_blank')),
+                'compatible' => html_writer::link('http://www.gnu.org/licenses/license-list.en.html#apache2', 'compatible',
+                    array('target' => '_blank')),
+                'contentlicenseurl' => html_writer::link('http://creativecommons.org/licenses/by/3.0/', 'CC BY 3.0',
+                    array('target' => '_blank')),
+                'globalsettings' => html_writer::link('http://getbootstrap.com/2.3.2/scaffolding.html#global', 'Global settings',
+                    array('target' => '_blank'))
+            )
+        )
+    ));
+    $ADMIN->add('theme_shoelace', $temp);
 }